1. id исправил на Long 2. equals+hash 3. @Transactional оставил только в UserServiceImpl 4. убрал репозиторий 5. добавил редактирование юзера

+++ src/main/java/spring/controller/MainController.java
@@ -3,7 +3,6 @@ package spring.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 import spring.service.UserService;
 
@@ -20,7 +19,7 @@ public class MainController {
     }
 
     @GetMapping("/show")
-    private String deleteUser( int id ) {
+    private String deleteUser( Long id ) {
         userService.delete(id);
         return "redirect:/";
     }
